Удаление ненужных файлов
